Add 5 year water mask (#11)
* Correction to the sensor cloud mask values

oli values were not correct

* Reversing previous commit

QA values for oli

* Update indexing step for instrument data

* Set the OLI QA threshold to <= 21952

21952 is the qa value for water, cloud flags set to low. Whilst the documentations suggests 21888, that overlooks that the pixels have the 'clear' flag set, as well as the 'water' flag, taking the value to 21952

* Update odc-stats dependency

* Incoporate steps in prepare_config into config module

* Format files

* Add odc-loader dependency

* Pin odc-loader to install from repo

* Pin odc-stats

* Add custom cell magics module

* Update msi and msi_agm instrument bands

* FAI workbook added

This is a self-contained notebook to assess the area of water affeected by algae and weeds, using the FAI - floating algal index. It is multi-sensor and aims to produce real information.

* Set the default area of interest

* deleted personal working notebook

* Revised the handling of FAI

Included a time-series graph
Introduced a results dictionary/dataframe.

* fixed the geomedian ndvi function

-introduced a water mask as an argument
-addressed code error for average ndvi across multiple instruments
-removed redundant function 'geomedian_NDVI_percent_affected'

* Implemented a water mask for NDVI

* Set a Threshold of 0.05 on NDVI values.

Previous threshold was 0.0. This is a parameter that we migth refine further in future

* Introduced a boolean variable  'clearwater'

Clearwater is defined as:
water_5yr ==1, fai_mask == False
water quality variables will only be valid when measured over clear water.

* Threshold on ndvi set to 0.05 rather than zero

This threshold should probably be an argument to the function, but later...

* Minor edits of typos

* New hue calculation functions handle all sensors

Re-wrote the hue_calculation and hue_adjust functions to be able to handle all sensors.
Removed internal references to wofs. Better to assume that the data paseed in will be limited to the area of interest with a mask.
Old code temporarily still there for visibility.

* Added a cell to calculate the hue

* Update test areas

* Add lower limit for datacube

* Update how wofs_ann data is loaded

Load 5 years worth of wofs_annual data and generate water mask

* revised Hue calculation function tested

The functions now work on multiple sensors.

* Revised geomedian hue calculation

function geomedian_hue introduced to calculate the hue values for each instrument in the multivariate geomedian dataset.

* updates to r_correction and hue_calculation

Revised hue calculation, added a function to calcuate the hue for geomedians: geomedian_hue. These now take a water mask rather than internal thresholds..
These will now work for dark-pixel-corrected geomedian data, as well as for non-geomedian data.
Revised R_correction and added 'apply_R_correction' to handle data preparation / contingencies. Added use of a mask for water threholding - nominally the clearwater mask.

* Add pre-commit checks

* Remove  jupyter checkpoints dir

* Update odc-loader pin

* Do not edit original query object

* OWT functions re-write

These functions are more transparent, don't kill the memory, cater for tm, oli and msi data, make use of all available wavelengths and are flexible for geomedians or other.
Might still have bugs; I need to run a very small test dataset through to validate.

* Updated

This notebook is updated for the steps up to calculation of the water quality measures, e.g., the calculation of FAI and NDVI, hue and OWT, the dark-pixel correction, new 'clearwater' mask.

* reference data  - sensor bands in a csv file

used in calculation of OWT

* Optical water type reference spectra

This is the table of spectral reflectance curves for each open water type.
Used in calculation of the OWT.

* Update load_wofs_ann

* Update loading 5 year water mask

* Fix data type

* Add seperate functions to load data for different instruments

* Fix high memory consumption in load_tirs_annual_composite_data

* minor fix

* minor fix

* Remove repeated log statement

* Rename watermask in water_analysis

* Water mask dtype to float32

* minor edit

* Update continental workflow to include 5year water mask

---------

Co-authored-by: alewis_01 <adam.lewis@digitalearthafrica.org>