Added new wq indices / functions to test. Dictionary-setting now here as a function
The dictionary for controlling algorithms / sensors is now a function in here. Simplified to support data from geomedians and other.
the 'r' suffix on vatriables is no longer 'supported'.  No point in retaining copies of the data pre and post dark pixel correction.
New functions are Tebbs, 3BDA, and Toming.